+++ README.md
@@ -104,6 +104,7 @@
         - [x]  Delete: 저장된 모든 바우처 삭제
 
 - [ ] 콘솔로 CRUD 실행 구현
+- [ ] 트랜잭션 문제 해결
 
 ## (심화) 바우처 지갑을 만들어보세요.
