+++ src/main/java/prgms/vouchermanagementapp/domain/FileVoucherRecord.java
@@ -1,0 +1,19 @@
+package prgms.vouchermanagementapp.domain;
+
+public class FileVoucherRecord implements VoucherRecord {
+
+    private final String filePath;
+
+    public FileVoucherRecord(String filePath) {
+        this.filePath = filePath;
+    }
+
+    public String getFilePath() {
+        return filePath;
+    }
+
+    @Override
+    public boolean isFile() {
+        return true;
+    }
+}
