+++ src/main/java/prgms/vouchermanagementapp/domain/VoucherRecord.java
@@ -1,0 +1,6 @@
+package prgms.vouchermanagementapp.domain;
+
+public interface VoucherRecord {
+
+    boolean isFile();
+}
