--- src/main/java/prgms/vouchermanagementapp/io/Reader.java
@@ -1,25 +1,0 @@
-package prgms.vouchermanagementapp.io;
-
-import org.springframework.stereotype.Component;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-@Component
-public class Reader {
-
-    private final BufferedReader bufferedReader;
-
-    public Reader() {
-        this.bufferedReader = new BufferedReader(new InputStreamReader(System.in));
-    }
-
-    public String readLine() {
-        try {
-            return bufferedReader.readLine();
-        } catch (IOException exception) {
-            throw new IllegalArgumentException(exception);
-        }
-    }
-}
