+++ src/main/java/prgms/vouchermanagementapp/service/VoucherManager.java
@@ -1,0 +1,32 @@
+package prgms.vouchermanagementapp.service;
+
+import org.springframework.stereotype.Component;
+import prgms.vouchermanagementapp.domain.Voucher;
+import prgms.vouchermanagementapp.domain.VoucherRecord;
+import prgms.vouchermanagementapp.domain.value.Amount;
+import prgms.vouchermanagementapp.domain.value.Ratio;
+import prgms.vouchermanagementapp.repository.Vouchers;
+
+@Component
+public class VoucherManager {
+
+    private final Vouchers vouchers;
+
+    public VoucherManager(Vouchers vouchers) {
+        this.vouchers = vouchers;
+    }
+
+    public void createVoucher(Amount fixedDiscountAmount) {
+        Voucher fixedAmountVoucher = VoucherCreationFactory.createVoucher(fixedDiscountAmount);
+        vouchers.store(fixedAmountVoucher);
+    }
+
+    public void createVoucher(Ratio fixedDiscountRatio) {
+        Voucher percentDiscountVoucher = VoucherCreationFactory.createVoucher(fixedDiscountRatio);
+        vouchers.store(percentDiscountVoucher);
+    }
+
+    public VoucherRecord findAllVouchers() {
+        return vouchers.getVoucherRecord();
+    }
+}
