--- src/main/java/prgms/vouchermanagementapp/storage/model/FileVoucherRecord.java
@@ -1,19 +1,0 @@
-package prgms.vouchermanagementapp.storage.model;
-
-public class FileVoucherRecord implements VoucherRecord {
-
-    private final String filePath;
-
-    public FileVoucherRecord(String filePath) {
-        this.filePath = filePath;
-    }
-
-    public String getFilePath() {
-        return filePath;
-    }
-
-    @Override
-    public boolean isFile() {
-        return true;
-    }
-}
