--- src/main/java/prgms/vouchermanagementapp/storage/model/MemoryVoucherRecord.java
@@ -1,23 +1,0 @@
-package prgms.vouchermanagementapp.storage.model;
-
-import prgms.vouchermanagementapp.domain.Voucher;
-
-import java.util.List;
-
-public class MemoryVoucherRecord implements VoucherRecord {
-
-    private final List<Voucher> memoryVouchers;
-
-    public MemoryVoucherRecord(List<Voucher> memoryVouchers) {
-        this.memoryVouchers = memoryVouchers;
-    }
-
-    @Override
-    public boolean isFile() {
-        return false;
-    }
-
-    public List<Voucher> getMemoryVouchers() {
-        return this.memoryVouchers;
-    }
-}
