--- src/main/java/prgms/vouchermanagementapp/storage/model/VoucherRecord.java
@@ -1,6 +1,0 @@
-package prgms.vouchermanagementapp.storage.model;
-
-public interface VoucherRecord {
-
-    boolean isFile();
-}
